Renamed slice folders
